jjjjjjj

+++ src/orders/models.py
@@ -27,7 +27,7 @@ class Order(models.Model):
     status = models.CharField(max_length=255,choices=Status_choices,default='confirmed')
     delivery_address = models.ForeignKey(Address, on_delete=models.CASCADE,null=True,default=None)
     payment_method = models.CharField(max_length=255,choices=payment_options,default='cash_on_delivery')
-    total_price = models.DecimalField(max_digits=11, decimal_places=2, default=0)
+    total_price = models.FloatField(default=0.0)
 
     def __str__(self):
         return f"{self.buyer}"
